+++ projet/phase1/read_stsp.jl
@@ -93,8 +93,8 @@ end
 """Analyse un fichier .tsp et renvoie l'ensemble des arêtes sous la forme d'un tableau."""
 function read_edges(header::Dict{String}{String}, filename::String)
 
-  edges = Tuple{Int64, Int64}[]
-  weights = Int64[]
+  edges = Tuple{Float64, Float64}[]
+  weights = Float64[]
   edge_weight_format = header["EDGE_WEIGHT_FORMAT"]
   known_edge_weight_formats = ["FULL_MATRIX", "UPPER_ROW", "LOWER_ROW",
   "UPPER_DIAG_ROW", "LOWER_DIAG_ROW", "UPPER_COL", "LOWER_COL",
